app/requirements.txt
tornado
pymysql
requests
pika
time_uuid
redis
pyyaml
sqlalchemy
clickhouse-driver
kafka-python
sqlalchemy-clickhouse